requirements.txt
scikit-learn
mypy
tqdm
ipykernel
spotipy
python-dotenv
pandas
types-requests
matplotlib
ipywidgets
seaborn
torch
numpy
librosa
flask
requests
pytube
# pyaudio
scikit-image
pychorus
kornia
torchaudio
ffmpeg-python
